pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "vggt-boxer-pipeline"
version = "0.1.0"
description = "iPhone video → 3D scene graph via VGGT + Boxer (no LiDAR required)"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    "torch>=2.3",
    "torchvision>=0.18",
    "numpy>=1.26",
    "Pillow",
    "tqdm",
]

[project.optional-dependencies]
viz = ["open3d", "matplotlib"]

[tool.setuptools.packages.find]
include = ["pipeline*"]
